Edit jest tests according to the backend changes

+++ backend/src/tests/testUtils.ts
@@ -1,12 +1,16 @@
 import request from "supertest";
 import { Express } from "express";
 import User from "../models/userModel";
+import Post from "../models/postModel";
+import Comment from "../models/commentModel";
+import Like from "../models/likeModel";
+import Save from "../models/saveModel";
 
 export interface TestUserData {
   username: string;
   email: string;
   password: string;
-  token?: string;
+  accessToken?: string;
   refreshToken?: string;
 }
 
@@ -14,7 +18,7 @@ export interface TestUser {
   _id: string;
   username: string;
   email: string;
-  token: string;
+  accessToken: string;
   refreshToken: string;
 }
 
@@ -38,14 +42,41 @@ export const registerTestUser = async (app: Express): Promise<TestUser> => {
   }
 
   return {
-    _id: response.body._id || "testUserId",
+    _id: response.body.user?._id || response.body._id || "testUserId",
     username: userData.username,
     email: userData.email,
-    token: response.body.token,
+    accessToken: response.body.accessToken,
     refreshToken: response.body.refreshToken,
   };
 };
 
 export const cleanupDatabase = async (): Promise<void> => {
-  await User.deleteMany({});
+  // Cleanup test created data
+  const testUsers = await User.find(
+    { email: /@example\.com$/i },
+    { _id: 1 },
+  ).lean();
+
+  if (testUsers.length === 0) {
+    return;
+  }
+
+  const userIds = testUsers.map((u) => u._id);
+  const testPosts = await Post.find(
+    { user: { $in: userIds } },
+    { _id: 1 },
+  ).lean();
+  const postIds = testPosts.map((p) => p._id);
+
+  await Like.deleteMany({
+    $or: [{ userId: { $in: userIds } }, { postId: { $in: postIds } }],
+  });
+  await Save.deleteMany({
+    $or: [{ userId: { $in: userIds } }, { postId: { $in: postIds } }],
+  });
+  await Comment.deleteMany({
+    $or: [{ user: { $in: userIds } }, { post: { $in: postIds } }],
+  });
+  await Post.deleteMany({ user: { $in: userIds } });
+  await User.deleteMany({ _id: { $in: userIds } });
 };
